Refactor around ListenableState and gen'd Interface
