+++ bin/server.dart
@@ -2,20 +2,19 @@
 
 import 'package:shelf/shelf.dart';
 import 'package:shelf/shelf_io.dart' as shelf_io;
-import "package:shelf_route/shelf_route.dart" as routing;
+import "package:shelf_router/shelf_router.dart" as routing;
 
 void main() async {
   var handler =
       const Pipeline().addMiddleware(logRequests()).addHandler(_echoRequest);
 
-  var server = await shelf_io.serve(handler, 'localhost', 8080);
+  var router = routing.Router();
+  router.get("/events", handler);
+
+  var server = await shelf_io.serve(router, 'localhost', 8080);
 
   // Enable content compression
   server.autoCompress = true;
-
-  var router = routing.router();
-  router.get("/events", handler);
-  router.get("/mychannel", channelHandler);
 
   print('Serving at http://${server.address.host}:${server.port}');
 }
