+++ lib/gen/interface.dart
@@ -1,7 +1,7 @@
-import 'shared.dart';
-export 'shared.dart';
-
 import 'package:flutter/foundation.dart';
+
+import 'package:counter_fullstack/infra/shared.dart';
+import 'package:counter_fullstack/backend.dart' as backend;
 
 // Is this just a https://api.flutter.dev/flutter/widgets/AsyncSnapshot-class.html?
 // This differs from ValueNotifier, in that its read-only from the consumer.
@@ -30,31 +30,34 @@ class CachedValue<T> extends ChangeNotifier implements ValueListenable<T> {
   String toString() => '$value';
 }
 
-class Connection {
-  Connection();
+// generated code
+class MyInterface {
+  MyInterface();
 
-  // What happens if this is called twice?
-  // We really want to subscribe to the stream from the server in a
-  // multi-cast fashion, no?
-  CachedValue<T> cache<T>(Stream<T> Function(Session) getter,
-      {required T initial}) {
+  static const int initialCount = 0;
+
+// What happens if this is called twice?
+// We really want to subscribe to the stream from the server in a
+// multi-cast fashion, no?
+  CachedValue<int> getCount() {
+    // Do all the work to call across the wire.
+
     // turn the function pointer into a url
     // Set the initial value to the passed initial value.
-    var value = CachedValue<T>(initial);
+    var value = CachedValue<int>(initialCount);
     // call the url
     // Sign up for the stream, update the cached value stream updates.
-    getter(Session()).listen((event) {
+    backend.getCount(Session()).listen((event) {
       value._updateValue(event);
     });
     return value;
   }
 
-  void call(void Function(Session) function) {
+  void incrementCounter() {
     // Turn the function pointer into a url
     // call to the url
     // Do I wait for the response?
-
-    // HACK for now: just call the function
-    function(Session());
+    // Hack, just call the function directly.
+    backend.incrementCount(Session());
   }
 }
